fix rune page upload
