+++ plugins/utils.js
@@ -1,5 +1,4 @@
 const rp = require('request-promise-native');
-const { groupBy } = require('lodash');
 
 /**
  * Gets a JSON object from a URI via 'request-promise-native'.
@@ -43,67 +42,14 @@ function getPerksMap(usedKeyStr = 'name') {
 }
 
 /**
- * Sorts the runes as they are in the game.
+ * Filters the perk ids from a list of runes ids.
  * @param {Array} runes - A list of runes ids
- * @return {Array} Sorted list of runes
+ * @return {Array} A list of runes ids without the ids of the perks.
  */
-function sortRunes(runes) {
-  // Index map for later sorting
-  const indexes = new Map();
-
-  // Sorting function that sorts the runes based on the index in the tree
-  const sortingFunc = (a, b) => indexes.get(a) - indexes.get(b);
-
-  // Creates a tree of style id and the matching runes
-  const tree = freezer.get().runesreforgedinfo.reduce((obj, curr) => {
-    obj[curr.id] = [].concat(...curr.slots.map(row => row.runes.map(i => i.id)));
-    return obj;
-  }, {});
-
-  // Creates a list of style ids based on the tree
-  const styleIds = Object.keys(tree).map(Number);
-
-  // Filters style ids from the runes
-  const filteredRunes = runes.filter(rune => !styleIds.includes(rune));
-
-  // Groups the passed runes fit to the respective style id
-  const groupedRunes = groupBy(filteredRunes, (rune) => {
-    for (const style of styleIds) {
-      const runeIndex = tree[style].indexOf(rune);
-      if (runeIndex !== -1) {
-        indexes.set(rune, runeIndex);
-        return style;
-      }
-    }
-  });
-
-  // Variables for determining the 'primaryStyleId' and 'secondaryStyleId
-  let primaryStyleLength = -1;
-  let primaryStyleId = -1;
-  let secondaryStyleLength = -1;
-  let secondaryStyleId = -1;
-
-  // Get 'primaryStyleId' and 'secondaryStyleId' based on the number of runes
-  for (const styleId in groupedRunes) {
-    if(styleId !== 'undefined'){
-      if(groupedRunes[styleId].length > primaryStyleLength){
-        secondaryStyleId = primaryStyleId;
-        primaryStyleId = styleId;
-        primaryStyleLength = groupedRunes[styleId].length;
-      }else if(groupedRunes[styleId].length >= secondaryStyleLength){
-        secondaryStyleId = styleId;
-        secondaryStyleLength = groupedRunes[styleId].length;
-      }
-    }
-  }
-
-  // Sorts the groups for the respective style
-  groupedRunes[primaryStyleId].sort(sortingFunc);
-  groupedRunes[secondaryStyleId].sort(sortingFunc);
-
-  // Merges primary and secondary runes in the correct order
-  return groupedRunes[primaryStyleId].concat(groupedRunes[secondaryStyleId], groupedRunes['undefined'] ?? []);
+function removePerkIds(runes) {
+  const perkIds = freezer.get().runesreforgedinfo.map(runeInfo =>  runeInfo.id);
+  return runes.filter(el => !perkIds.includes(el));
 }
 
 // Transfer of functions for use in other modules
-module.exports = { getStylesMap, getPerksMap, getJson, sortRunes };
+module.exports = { getStylesMap, getPerksMap, getJson, removePerkIds };
